fix: Fix code repository option no init when after search in allowlist
